added test for operationMultiEval
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@4641 4eea3252-f0fb-4393-894d-40516dce545b

+++ base/tests/test_DataVector.cpp
@@ -116,7 +116,7 @@ BOOST_AUTO_TEST_CASE(testDotProduct) {
   DataVector d = DataVector(3);
   SGPP::float_t x = 0;
 
-  for (int i = 0; i < static_cast<int>(d.getSize()); ++i) {
+  for (unsigned int i = 0; i < d.getSize(); ++i) {
     d[i] = i + 1;
     x += d[i] * d[i];
   }
